Adding blob sign command
